echoAI-Production/echolib/Get_connector/Get_API/requirements.txt
httpx==0.27.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
cryptography==42.0.0
pyjwt==2.9.0
python-dotenv==1.0.0
